# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeExplainer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AnalysisLevel = typeof AnalysisLevel[keyof typeof AnalysisLevel];


export const AnalysisLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Analysis {
  id: number;
  title: string;
  githubUrl: string;
  level: AnalysisLevel;
  status: AnalysisStatus;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  architecture?: string | null;
  /** @nullable */
  keyFunctions?: string | null;
  /** @nullable */
  dataFlow?: string | null;
  /** @nullable */
  walkthrough?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

/**
 * Explanation detail level
 */
export type AnalysisInputLevel = typeof AnalysisInputLevel[keyof typeof AnalysisInputLevel];


export const AnalysisInputLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface AnalysisInput {
  /** GitHub repository URL (e.g. https://github.com/owner/repo) */
  githubUrl: string;
  /** Explanation detail level */
  level: AnalysisInputLevel;
}

export type AnalysisStatsByLevel = {
  beginner: number;
  intermediate: number;
  advanced: number;
};

export interface AnalysisStats {
  total: number;
  completed: number;
  pending: number;
  byLevel: AnalysisStatsByLevel;
  recentAnalyses: Analysis[];
}

export interface ApiError {
  error: string;
}