# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeExplainer API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisLevel } from './analysisLevel';
import type { AnalysisStatus } from './analysisStatus';

export interface Analysis {
  id: number;
  title: string;
  githubUrl: string;
  level: AnalysisLevel;
  status: AnalysisStatus;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  architecture?: string | null;
  /** @nullable */
  keyFunctions?: string | null;
  /** @nullable */
  dataFlow?: string | null;
  /** @nullable */
  walkthrough?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: Date;
  updatedAt: Date;
}