# lib/api-zod/src/generated/types/analysisInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeExplainer API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisInputLevel } from './analysisInputLevel';

export interface AnalysisInput {
  /** GitHub repository URL (e.g. https://github.com/owner/repo) */
  githubUrl: string;
  /** Explanation detail level */
  level: AnalysisInputLevel;
}