# lib/api-zod/src/generated/types/analysisStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeExplainer API
 * OpenAPI spec version: 0.1.0
 */
import type { Analysis } from './analysis';
import type { AnalysisStatsByLevel } from './analysisStatsByLevel';

export interface AnalysisStats {
  total: number;
  completed: number;
  pending: number;
  byLevel: AnalysisStatsByLevel;
  recentAnalyses: Analysis[];
}